requirements.txt
pytorch==1.8.0
torchvision==0.9.0
matplotlib